misc tidy-ups
